+++ lib/settings_reader/vault_resolver/engines/abstract.rb
@@ -16,7 +16,7 @@ module SettingsReader
         end
 
         def get(address)
-          return unless (vault_secret = get_secret(address))
+          return unless (vault_secret = get_secret_with_retries(address))
 
           wrap_secret(address, vault_secret)
         rescue Vault::VaultError => e
@@ -26,7 +26,7 @@ module SettingsReader
         def renew(entry)
           return unless entry.leased?
 
-          new_secret = renew_lease(entry)
+          new_secret = renew_lease_with_retries(entry)
           entry.update_renewed(new_secret)
           true
         rescue Vault::VaultError => e
@@ -34,6 +34,18 @@ module SettingsReader
         end
 
         protected
+
+        def get_secret_with_retries(address)
+          Vault.with_retries(Vault::HTTPConnectionError, attempts: config.retrieval_retries) do
+            get_secret(address)
+          end
+        end
+
+        def renew_lease_with_retries(address)
+          Vault.with_retries(Vault::HTTPConnectionError, attempts: config.lease_renew_retries) do
+            renew_lease(address)
+          end
+        end
 
         def get_secret(address)
           raise NotImplementedError